hw4/requirements.txt
cupy-cuda102==11.3.0
Pillow==9.2.0
